requirements.txt
Django>=5.1.3,<6
django-cryptography==1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10